refactor: update getCompanyById controller function
