+++ structs/validateUuid.js
